# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api.ErrorResponse:
    description: Error response structure
    properties:
      code:
        example: INVALID_INPUT
        type: string
      error:
        example: bad_request
        type: string
      message:
        example: Invalid request parameters
        type: string
    type: object
...
  api.ManifestFileResponse:
    properties:
      hash:
        description: SHA256 hash
        type: string
      path:
        type: string
      size:
        type: integer
    type: object
  api.ManifestResponse:
    properties:
      app:
        type: string
      build_time:
        type: string
      builder:
        type: string
      files:
        items:
          $ref: '#/definitions/api.ManifestFileResponse'
        type: array
      git_commit:
        type: string
      hash:
        description: Same as version
        type: string
      project:
        type: string
      version:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: kk
    url: https://opensource.org/licenses/MIT
  description: |-
    kkArtifact is a modern artifact management and synchronization system designed to replace rsync + SSH deployment workflows.
    The system provides multi-project and multi-app artifact storage with hash-based versioning, HTTP-based API for artifact upload, download, and management.
    Token-based authentication with fine-grained permissions (Global/Project/App scopes).
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: kkArtifact API
  version: "1.0"
paths:
...
  /manifest/{project}/{app}/{hash}:
    get:
      consumes:
      - application/json
      description: Get the manifest for a specific version (includes file list, metadata,
        Git commit info)
      parameters:
      - description: Project name
        in: path
        name: project
        required: true
        type: string
      - description: App name
        in: path
        name: app
        required: true
        type: string
      - description: Version hash
        in: path
        name: hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ManifestResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - Bearer: []
      summary: Get manifest
      tags:
      - artifacts
...
securityDefinitions:
  Bearer:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"